requirements.txt
absl-py==0.12.0
amqp==5.0.6
asgiref==3.3.4
astunparse==1.6.3
billiard==3.6.4.0
cachetools==4.2.2
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
decorator==4.4.2
Django==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.12.4
flatbuffers==1.12
gast==0.4.0
gcloud==0.17.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.34.1
gunicorn==20.1.0
h5py==3.1.0
httplib2==0.19.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
jws==0.1.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kombu==5.0.2
Markdown==3.3.4
moviepy==1.0.3
numpy==1.19.5
oauth2client==3.0.0
oauthlib==3.1.0
opencv-python==4.5.2.52
opt-einsum==3.3.0
Pillow==8.2.0
proglog==0.1.9
prompt-toolkit==3.0.18
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
pyparsing==2.4.7
Pyrebase==3.0.27
python-jwt==2.0.1
pytz==2021.1
requests==2.11.1
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
rsa==4.7.2
six==1.15.0
SpeechRecognition==3.8.1
sqlparse==0.4.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.0
wrapt==1.12.1
